Add value reader

+++ corewar/src/instruction/special_param/zjmp.c
@@ -10,6 +10,6 @@
 int special_param_zjmp(parameter_t *parameter, int index)
 {
     if (index == 0)
-        parameter->type = Index;
+        parameter->type = (short)Index;
     return 0;
 }
